perf(core): Lazyload security audit reporters
