+++ packages/cli/src/PublicApi/types.ts
@@ -9,7 +9,7 @@ import type { WorkflowEntity } from '@db/entities/WorkflowEntity';
 
 import type { UserManagementMailer } from '@/UserManagement/email';
 
-import type { Risk } from '@/audit/types';
+import type { Risk } from '@/security-audit/types';
 
 export type AuthlessRequest<
 	RouteParams = {},
