+++ packages/cli/src/commands/audit.ts
@@ -1,8 +1,8 @@
 import { flags } from '@oclif/command';
-import { audit } from '@/audit';
-import { RISK_CATEGORIES } from '@/audit/constants';
+import { SecurityAuditService } from '@/security-audit/SecurityAudit.service';
+import { RISK_CATEGORIES } from '@/security-audit/constants';
 import config from '@/config';
-import type { Risk } from '@/audit/types';
+import type { Risk } from '@/security-audit/types';
 import { BaseCommand } from './BaseCommand';
 import { Container } from 'typedi';
 import { InternalHooks } from '@/InternalHooks';
@@ -49,7 +49,10 @@ export class SecurityAudit extends BaseCommand {
 			throw new Error([message, hint].join('. '));
 		}
 
-		const result = await audit(categories, auditFlags['days-abandoned-workflow']);
+		const result = await Container.get(SecurityAuditService).run(
+			categories,
+			auditFlags['days-abandoned-workflow'],
+		);
 
 		if (Array.isArray(result) && result.length === 0) {
 			this.logger.info('No security issues found');
