+++ packages/cli/test/integration/security-audit/DatabaseRiskReporter.test.ts
@@ -1,18 +1,22 @@
 import { v4 as uuid } from 'uuid';
-import { audit } from '@/audit';
+import { SecurityAuditService } from '@/security-audit/SecurityAudit.service';
 import {
 	DATABASE_REPORT,
 	SQL_NODE_TYPES,
 	SQL_NODE_TYPES_WITH_QUERY_PARAMS,
-} from '@/audit/constants';
+} from '@/security-audit/constants';
 import { getRiskSection, saveManualTriggerWorkflow } from './utils';
 import * as testDb from '../shared/testDb';
 import { generateNanoId } from '@db/utils/generators';
 import { WorkflowRepository } from '@db/repositories/workflow.repository';
 import Container from 'typedi';
 
+let securityAuditService: SecurityAuditService;
+
 beforeAll(async () => {
 	await testDb.init();
+
+	securityAuditService = new SecurityAuditService(Container.get(WorkflowRepository));
 });
 
 beforeEach(async () => {
@@ -56,7 +60,7 @@ test('should report expressions in queries', async () => {
 
 	await Promise.all(promises);
 
-	const testAudit = await audit(['database']);
+	const testAudit = await securityAuditService.run(['database']);
 
 	const section = getRiskSection(
 		testAudit,
@@ -111,7 +115,7 @@ test('should report expressions in query params', async () => {
 
 	await Promise.all(promises);
 
-	const testAudit = await audit(['database']);
+	const testAudit = await securityAuditService.run(['database']);
 
 	const section = getRiskSection(
 		testAudit,
@@ -163,7 +167,7 @@ test('should report unused query params', async () => {
 
 	await Promise.all(promises);
 
-	const testAudit = await audit(['database']);
+	const testAudit = await securityAuditService.run(['database']);
 
 	const section = getRiskSection(
 		testAudit,
@@ -183,7 +187,7 @@ test('should report unused query params', async () => {
 test('should not report non-database node', async () => {
 	await saveManualTriggerWorkflow();
 
-	const testAudit = await audit(['database']);
+	const testAudit = await securityAuditService.run(['database']);
 
 	expect(testAudit).toBeEmptyArray();
 });
